# services/price-service/docs/swagger.yaml
basePath: /internal
definitions:
...
  handlers.IngestionStoreStats:
    properties:
      failedRows:
        type: integer
      fileCount:
        type: integer
      persistedCount:
        type: integer
      priceChanges:
        type: integer
      rowCount:
        type: integer
      storeCity:
        type: string
      storeId:
        type: string
      storeIdentifier:
        type: string
      storeName:
        type: string
      warningRows:
        type: integer
    type: object
...
  handlers.ListStoreStatsResponse:
    properties:
      stores:
        items:
          $ref: '#/definitions/handlers.IngestionStoreStats'
        type: array
      total:
        type: integer
    type: object
...
info:
  contact: {}
  description: Internal API for price data management, ingestion monitoring, and basket
    optimization.
  title: Price Service API
  version: "1.0"
paths:
...
  /internal/ingestion/files/{fileId}/stores:
    get:
      consumes:
      - application/json
      description: Returns store stats for a specific ingestion file
      parameters:
      - description: File ID
        in: path
        name: fileId
        required: true
        type: string
      - default: 50
        description: Number of items to return
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Number of items to skip
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ListStoreStatsResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List ingestion store stats (file)
      tags:
      - ingestion
...
  /internal/ingestion/runs/{runId}/stores:
    get:
      consumes:
      - application/json
      description: Returns aggregated store stats for a specific ingestion run
      parameters:
      - description: Run ID
        in: path
        name: runId
        required: true
        type: string
      - default: 50
        description: Number of items to return
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Number of items to skip
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ListStoreStatsResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List ingestion store stats (run)
      tags:
      - ingestion
...
swagger: "2.0"